refactor: introduce PresentableRouter and PresentationModel
Extract sheet/cover/close presentation logic into a dedicated
PresentableRouter class between BaseRouter and Router. TabRouter
stays clean. Router and future SplitRouter inherit presentation
for free. PresentationModel defines the public protocol contract.
